requirements.txt
python-telegram-bot==20.8
yfinance==0.2.36
python-dotenv==1.0.1
